+++ tests/Serializer/CollectionDenormalizerTest.php
@@ -5,10 +5,10 @@ namespace Mtarld\ApiPlatformMsBundle\Tests\Serializer;
 use ApiPlatform\Core\DataProvider\ArrayPaginator;
 use Mtarld\ApiPlatformMsBundle\Collection\Collection;
 use Mtarld\ApiPlatformMsBundle\Collection\Pagination;
-use Mtarld\ApiPlatformMsBundle\Tests\BcLayerKernelTestCase;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Dto\PuppyDto;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Dto\PuppyResourceDto;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Entity\Puppy;
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
@@ -17,7 +17,7 @@ use Symfony\Component\Serializer\SerializerInterface;
  *
  * @author Mathias Arlaud <mathias.arlaud@gmail.com>
  */
-class CollectionDenormalizerTest extends BcLayerKernelTestCase
+class CollectionDenormalizerTest extends KernelTestCase
 {
     public function setUp(): void
     {