+++ tests/Serializer/ObjectDenormalizerTest.php
@@ -2,9 +2,9 @@
 
 namespace Mtarld\ApiPlatformMsBundle\Tests\Serializer;
 
-use Mtarld\ApiPlatformMsBundle\Tests\BcLayerKernelTestCase;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Dto\PuppyDto;
 use Mtarld\ApiPlatformMsBundle\Tests\Fixtures\App\src\Entity\Puppy;
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
@@ -13,7 +13,7 @@ use Symfony\Component\Serializer\SerializerInterface;
  *
  * @author Mathias Arlaud <mathias.arlaud@gmail.com>
  */
-class ObjectDenormalizerTest extends BcLayerKernelTestCase
+class ObjectDenormalizerTest extends KernelTestCase
 {
     public function setUp(): void
     {