+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Dropped API Platform < 2.6.4 support
+* Dropped Symfony < 5.4 support
+* Dropped PHP7.1 support
+* Added Symfony 6 support
 
 ## [0.4.0] - 2022-01-07
 * Added dispatching of a `RequestEvent` once a request is sent